// src/types/dnd5BackendTypes.d.ts
export type actionData = {};
export type dnd5SavingThrows = {
    str?: number;
    dex?: number;
    int?: number;
    wis?: number;
    con?: number;
    cha?: number;
}
export type dnd5Sense = {
    name: string;
    value: string;
}
export type dnd5SpellSlots = {
    0: string;
    1: number;
    2: number;
    3: number;
    4: number;
    5: number;
    6: number;
    7: number;
    8: number;
    9: number;
}